% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star.R
\name{star}
\alias{star}
\title{A fonction that transforms pvalues in stars ***}
\usage{
star(p.val)
}
\arguments{
\item{p.val}{the value of the pvalue.}
}
\value{
A character string with stars, dot or blank space.
}
\description{
A fonction that transforms pvalues in stars ***
}
\author{
Julien Bousquet (2021)
}
